Added support for querying by project.
Added support for querying for summaries using a project id. Per_sample() function had this functionality, but the calling functions above it did not support it.
A new helper function was added co-level to per_sample() to return the list of barcode ids given a project id. This allows us to have consistent behavior when querying using barcodes or with project id.