Modified summaries query
There isn't an API call to get a list of barcodes for a given project. I therefore modified the summaries query to return a partial answer if over 1000 barcodes are passed or more than 1000 barcodes are found in a project. The query was modified to return a dictionary containing a list of sample summaries under the 'samples' keyword. 'partial_result' is an additional key added to return "true" if the result is partial. It will be set to "false" otherwise. The remaining unprocessed barcodes are returned as a list under a third and final key 'unprocessed_barcodes'. This allows the user to ultimately retrieve all of the data if they wish by passing the remaining list in whole or in part to subsequent calls.