Fix test_edit_sample_locked()
Fixed test to use new existing survey.
relativedelta(month=2) appeared to have issues w/rollover for both month
and year when the current month is 11 (November). Replaced w/standard
datetime.timedelta() using 61 days, the median number of days in a given
two month span (from beginning of the month to the end of the next
month).