(WIP) migrate_responses_by_barcode() updated.
migrate_responses_by_barcode() now best fits the responses it returns
according to the values closest to a particular timestamp. This
timestamp can be defined separately in a helper method. Currently the
helper looks for a timestamp from a survey with the same template as the
one requested. If a survey doesn't exist, the method will return the
creation timestamp of the source associated w/the barcode.