Removed survey_template_id from timesstamp query
survey_template_id isn't needed for the revised get_timestamp() query,
as the timestamps for all results are going to be the same, due to the
way the query is written. Moreover, regardless of whether there are say
multiple template 10s, or no template 10s, we would want to center our
response results around this timestamp.